+++ packages/sdk/src/encryption/KeyExchangeKeyPair.ts
@@ -1,0 +1,7 @@
+import { AsymmetricEncryptionType } from '@streamr/trackerless-network'
+
+export interface KeyExchangeKeyPair {
+    getPublicKey(): Uint8Array
+    getPrivateKey(): Uint8Array
+    getEncryptionType(): AsymmetricEncryptionType
+}