+++ packages/sdk/src/encryption/SubscriberKeyExchange.ts
@@ -1,15 +1,12 @@
-import { Logger, StreamPartID, StreamPartIDUtils, UserID, toUserId } from '@streamr/utils'
+import { Logger, StreamPartID, StreamPartIDUtils, UserID, toUserId, toUserIdRaw } from '@streamr/utils'
 import { Lifecycle, delay, inject, scoped } from 'tsyringe'
 import { v4 as uuidv4 } from 'uuid'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
 import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
 import { NetworkNodeFacade } from '../NetworkNodeFacade'
 import { StreamRegistry } from '../contracts/StreamRegistry'
-import { GroupKeyRequest as OldGroupKeyRequest } from '../protocol/GroupKeyRequest'
-import { GroupKeyResponse as OldGroupKeyResponse } from '../protocol/GroupKeyResponse'
 import { MessageID } from '../protocol/MessageID'
 import { ContentType, EncryptionType, SignatureType, StreamMessage, StreamMessageType } from '../protocol/StreamMessage'
-import { convertBytesToGroupKeyResponse, convertGroupKeyRequestToBytes } from '../protocol/oldStreamMessageBinaryUtils'
 import { createRandomMsgChainId } from '../publish/messageChain'
 import { MessageSigner } from '../signature/MessageSigner'
 import { SignatureValidator } from '../signature/SignatureValidator'
@@ -18,9 +15,12 @@ import { LoggerFactory } from '../utils/LoggerFactory'
 import { pOnce, withThrottling } from '../utils/promises'
 import { MaxSizedSet } from '../utils/utils'
 import { validateStreamMessage } from '../utils/validateStreamMessage'
-import { GroupKey } from './GroupKey'
 import { LocalGroupKeyStore } from './LocalGroupKeyStore'
 import { RSAKeyPair } from './RSAKeyPair'
+import { EncryptionUtil } from './EncryptionUtil'
+import { MLKEMKeyPair } from './MLKEMKeyPair'
+import { AsymmetricEncryptionType, GroupKeyRequest, GroupKeyResponse } from '@streamr/trackerless-network'
+import { KeyExchangeKeyPair } from './KeyExchangeKeyPair'
 
 const MAX_PENDING_REQUEST_COUNT = 50000 // just some limit, we can tweak the number if needed
 
@@ -31,7 +31,7 @@ const MAX_PENDING_REQUEST_COUNT = 50000 // just some limit, we can tweak the num
 @scoped(Lifecycle.ContainerScoped)
 export class SubscriberKeyExchange {
 
-    private rsaKeyPair?: RSAKeyPair
+    private keyPair?: KeyExchangeKeyPair
     private readonly pendingRequests: MaxSizedSet<string> = new MaxSizedSet(MAX_PENDING_REQUEST_COUNT)
     private readonly networkNodeFacade: NetworkNodeFacade
     private readonly streamRegistry: StreamRegistry
@@ -64,7 +64,11 @@ export class SubscriberKeyExchange {
         this.authentication = authentication
         this.logger = loggerFactory.createLogger(module)
         this.ensureStarted = pOnce(async () => {
-            this.rsaKeyPair = await RSAKeyPair.create(config.encryption.rsaKeyLength)
+            if (config.encryption.requireQuantumResistantKeyExchange) {
+                this.keyPair = MLKEMKeyPair.create()
+            } else {
+                this.keyPair = await RSAKeyPair.create(config.encryption.rsaKeyLength)
+            }
             networkNodeFacade.addMessageListener((msg: StreamMessage) => this.onMessage(msg))
             this.logger.debug('Started')
         })
@@ -76,18 +80,18 @@ export class SubscriberKeyExchange {
     private async doRequestGroupKey(groupKeyId: string, publisherId: UserID, streamPartId: StreamPartID): Promise<void> {
         await this.ensureStarted()
         const requestId = uuidv4()
-        const request = await this.createRequest(
+        const { message, request } = await this.createRequest(
             groupKeyId,
             streamPartId,
             publisherId,
-            this.rsaKeyPair!.getPublicKey(),
             requestId)
-        await this.networkNodeFacade.broadcast(request)
+        await this.networkNodeFacade.broadcast(message)
         this.pendingRequests.add(requestId)
         this.logger.debug('Sent group key request (waiting for response)', {
             groupKeyId,
             requestId,
-            publisherId
+            publisherId,
+            keyEncryptionType: AsymmetricEncryptionType[request.encryptionType]
         })
     }
 
@@ -95,17 +99,17 @@ export class SubscriberKeyExchange {
         groupKeyId: string,
         streamPartId: StreamPartID,
         publisherId: UserID,
-        rsaPublicKey: string,
-        requestId: string
-    ): Promise<StreamMessage> {
-        const requestContent = new OldGroupKeyRequest({
-            recipient: publisherId,
+        requestId: string,
+    ): Promise<{ message: StreamMessage, request: GroupKeyRequest }> {
+        const request: GroupKeyRequest = {
+            recipientId: toUserIdRaw(publisherId),
             requestId,
-            rsaPublicKey,
+            publicKey: this.keyPair!.getPublicKey(),
             groupKeyIds: [groupKeyId],
-        })
+            encryptionType: this.keyPair!.getEncryptionType(),
+        }
         const erc1271contract = this.subscriber.getERC1271ContractAddress(streamPartId)
-        return this.messageSigner.createSignedMessage({
+        const message = await this.messageSigner.createSignedMessage({
             messageId: new MessageID(
                 StreamPartIDUtils.getStreamID(streamPartId),
                 StreamPartIDUtils.getStreamPartition(streamPartId),
@@ -114,24 +118,28 @@ export class SubscriberKeyExchange {
                 erc1271contract === undefined ? await this.authentication.getUserId() : toUserId(erc1271contract),
                 createRandomMsgChainId()
             ),
-            content: convertGroupKeyRequestToBytes(requestContent),
+            content: GroupKeyRequest.toBinary(request),
             contentType: ContentType.BINARY,
             messageType: StreamMessageType.GROUP_KEY_REQUEST,
             encryptionType: EncryptionType.NONE,
         }, erc1271contract === undefined ? SignatureType.SECP256K1 : SignatureType.ERC_1271)
+
+        return { message, request }
     }
 
     private async onMessage(msg: StreamMessage): Promise<void> {
-        if (OldGroupKeyResponse.is(msg)) {
+        if (msg.messageType === StreamMessageType.GROUP_KEY_RESPONSE) {
             try {
-                const { requestId, recipient, encryptedGroupKeys } = convertBytesToGroupKeyResponse(msg.content)
-                if (await this.isAssignedToMe(msg.getStreamPartID(), recipient, requestId)) {
+                const { requestId, recipientId, groupKeys: encryptedGroupKeys, encryptionType } = GroupKeyResponse.fromBinary(msg.content)
+                const recipientUserId = toUserId(recipientId)
+
+                if (await this.isAssignedToMe(msg.getStreamPartID(), recipientUserId, requestId)) {
                     this.logger.debug('Handle group key response', { requestId })
                     this.pendingRequests.delete(requestId)
                     await validateStreamMessage(msg, this.streamRegistry, this.signatureValidator)
                     await Promise.all(encryptedGroupKeys.map(async (encryptedKey) => {
-                        const key = GroupKey.decryptRSAEncrypted(encryptedKey, this.rsaKeyPair!.getPrivateKey())
-                        await this.store.set(key.id, msg.getPublisherId(), key.data)
+                        const key = await EncryptionUtil.decryptWithPrivateKey(encryptedKey.data, this.keyPair!.getPrivateKey(), encryptionType)
+                        await this.store.set(encryptedKey.id, msg.getPublisherId(), key)
                     }))
                 }
             } catch (err: any) {